requirements.txt
yt-dlp>=2024.0.0
opencv-python>=4.9.0
anthropic>=0.40.0
pandas>=2.2.0
openpyxl>=3.1.0
Pillow>=10.0.0
easyocr>=1.7.0
ultralytics>=8.0.0
